FIFA-World-Cup-Analysis
This file is part of the FIFA World Cup Analysis project, which aims to uncover and analyze the intricate patterns and hidden narratives behind football's most prestigious tournament. The file includes the necessary data and scripts to explore goal-scoring patterns of teams throughout the history of the FIFA World Cup.